+++ native/modules/combat_stats/module.cpp
@@ -6485,7 +6485,7 @@ void BE_CALL Shutdown() {
 }
 
 const BE_ModuleApiV1 kApi{
-    {kModuleId, "Combat Statistics", "3.0.1", BETTER_ENDFIELD_MODULE_ABI_V1},
+    {kModuleId, "Combat Statistics", "3.1.1", BETTER_ENDFIELD_MODULE_ABI_V1},
     &Initialize,
     &ConfigurationChanged,
     &Shutdown};