+++ native/modules/model/module.cpp
@@ -4996,7 +4996,7 @@ void BE_CALL Shutdown() {
 }
 
 const BE_ModuleApiV1 kApi{
-    {kModuleId, "Login Model", "3.0.1", BETTER_ENDFIELD_MODULE_ABI_V1},
+    {kModuleId, "Login Model", "3.1.1", BETTER_ENDFIELD_MODULE_ABI_V1},
     &Initialize,
     &ConfigurationChanged,
     &Shutdown};